fixed axios + Buffer dep
