+++ packages/anon-aadhaar-react/package.json
@@ -36,6 +36,7 @@
     "@types/uuid": "^9.0.2",
     "@types/valid-url": "^1.0.4",
     "anon-aadhaar-pcd": "*",
+    "buffer": "^6.0.3",
     "node-forge": "^1.3.1",
     "react-dom": "^18.2.0",
     "styled-components": "^6.0.5"
